Support sorting large lists by dividing and conquering on merge step of mergesort
This change generalizes the merge step of mergesort to what is being called "partial merge".

Partial merge algorithm receives two lists and a comparison function. Contrary to regular merge, it assumes there are extra elements after the input lists:

merge([e1, e2, e3, e4], [e5, e6, e7 e8], cmp) -> partial_merge([e1, e2, ?], [e5, e6, ?]) + partial_merge(...)

The algorithm can't compare an element to ?, therefore it returns two things:
1 - The partial result it was able to merge
2 - The "carry" which was whatever leftover of the inputs when it comes to compare elements with ?

In this algorithm, the two input lists are divided in two, and partial merge is applied to the first half of both lists.
The carry is appended to the second half of the input lists, and then partial merge is applied again.
The partial result of the invocation is the concatenation of the partial results of both partial merge recursive invocations.

The result of the partial merge invocation is the concatenation of the result of the two recursive invocations.
The carry is the carry of the second recursive invocation.

In this implementation, when the input lists are smaller than a given threshold, the algorithm falls back to the simpler and direct version of partial merge.
By doing so, there is a minimal cost of applying the divide/conquer for small lists. For large lists, the recursion depth will be limited, allowing the compiler to sort lists of thousands of elements.